requirements.txt
# AI & LLM
google-genai>=1.0.0
openai>=1.0.0
llama-index>=0.14.0
llama-index-llms-google-genai>=0.4.0
llama-index-llms-openai>=0.4.0
llama-index-embeddings-huggingface>=0.5.0

# Vector Store & Embeddings
chromadb>=1.3.0
llama-index-vector-stores-chroma>=0.4.0

# Agent Orchestration
langgraph>=1.0.0
langchain-core>=0.3.0

# Compute & Sandbox
modal>=1.2.0

# MCP Protocol
mcp>=1.22.0

# UI Framework
gradio>=6.0.0

# Database & ORM
sqlalchemy>=2.0.0
pymysql>=1.1.0

# Testing
pytest>=9.0.0
pytest-cov>=6.0.0
pytest-timeout>=2.3.0
pytest-asyncio>=0.24.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.10.0
transformers>=4.30.0  # For proper tokenization
